src/main/java/com/skypro/CourseWork2/JavaQuestionService.java
package com.skypro.CourseWork2;

import org.springframework.stereotype.Service;

import java.util.Collection;
import java.util.HashSet;
import java.util.Random;
import java.util.Set;

@Service
public class JavaQuestionService  implements QuestionService{

    Set<Question> questions = new HashSet<Question>();
    Random random = new Random();

    @Override
    public Question add(String question, String answer) {
        Question newQuestion = new Question(question,answer);
        questions.add(newQuestion);
        return newQuestion;
    }

    @Override
    public Question add(Question question) {
        questions.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        if (!questions.contains(question)) {
            throw new QuestionNotFoundException();
        }
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return questions;
    }

    @Override
    public Question getRandomQuestion() {
        Question [] questArr = questions.toArray(new Question[questions.size()]);
        return questArr[random.nextInt(questArr.length)];
    }
}
